Изменение лейаута, починка импортов
