+++ package.json
@@ -106,7 +106,7 @@
         "husky": "^7.0.0",
         "react-app-rewire-alias": "^1.0.3",
         "react-app-rewired": "^2.1.8",
-        "sass": "^1.35.2",
+        "sass": "^1.42.1",
         "stylelint": "^13.13.1",
         "stylelint-config-prettier": "^8.0.2",
         "stylelint-config-standard": "^22.0.0",
