+++ src/App.jsx
@@ -7,18 +7,18 @@ import '@fontsource/roboto/900.css'
 import '@fontsource/quantico'
 import '@fontsource/open-sans'
 import '@fontsource/open-sans/300.css'
-import '@/styles/index.scss'
+import '@/common/styles/index.scss'
 import Header from '@/components/Header/Header'
 import Landing from '@/pages/Landing/Landing'
 import NotFoundPage from '@/pages/NotFound/NotFound'
 import Footer from '@/components/Footer/Footer'
-import Login from '@/pages/Login/Login'
-import Signup from '@/pages/Signup/Signup'
+import Login from '@/pages/login'
+import Signup from '@/pages/signup'
 import Queue from '@/pages/Queue/Queue'
 import Privacy from '@/pages/Privacy/Privacy'
 import Feedback from '@/pages/Feedback/Feedback'
 import UserAgreement from '@/pages/UserAgreement/UserAgreement'
-import ghostRoutes from '@/routes/ghostRoutes'
+import ghostRoutes from '@/common/routes/ghostRoutes'
 import Metric from '@/utils/Metric'
 import ScrollToTop from '@/utils/ScrollToTop'
 import Technical from '@/pages/Technical/Technical'
