+++ src/common/components/Calendar/Calendar.module.scss
@@ -1,124 +1,122 @@
-@import '@/styles/variables';
+@import '@/common/styles/variables';
 
-:global {
-    // global отключает name mangling у scss. если его убрать всё сломается
+// global отключает name mangling у scss. если его убрать всё сломается
 
-    .react-calendar {
-        width: 100%;
-        font-size: 1.25rem;
-    }
-    .react-calendar,
-    .react-calendar *,
-    .react-calendar *::before,
-    .react-calendar *::after {
-        -moz-box-sizing: border-box;
-        -webkit-box-sizing: border-box;
-        box-sizing: border-box;
-    }
+.react-calendar {
+    width: 100%;
+    font-size: 1.25rem;
+}
+.react-calendar,
+.react-calendar *,
+.react-calendar *::before,
+.react-calendar *::after {
+    -moz-box-sizing: border-box;
+    -webkit-box-sizing: border-box;
+    box-sizing: border-box;
+}
 
-    .react-calendar button {
-        margin: 0;
-        outline: none;
-    }
+.react-calendar button {
+    margin: 0;
+    outline: none;
+}
 
-    .react-calendar__navigation {
-        height: 44px;
-        margin-bottom: 1em;
-    }
-    .react-calendar__navigation button {
-        min-width: 44px;
-        background: none;
-    }
-    .react-calendar button:enabled:hover {
-        cursor: pointer;
-    }
-    .react-calendar__navigation button:enabled:hover,
-    .react-calendar__navigation button:enabled:focus {
-        background-color: #e6e6e6;
-    }
-    .react-calendar__navigation button[disabled] {
-        background-color: #f0f0f0;
-    }
+.react-calendar__navigation {
+    height: 44px;
+    margin-bottom: 1em;
+}
+.react-calendar__navigation button {
+    min-width: 44px;
+    background: none;
+}
+.react-calendar button:enabled:hover {
+    cursor: pointer;
+}
+.react-calendar__navigation button:enabled:hover,
+.react-calendar__navigation button:enabled:focus {
+    background-color: #e6e6e6;
+}
+.react-calendar__navigation button[disabled] {
+    background-color: #f0f0f0;
+}
 
-    .react-calendar__month-view__weekdays {
-        text-align: center;
-        text-transform: capitalize;
-        font-weight: 400;
+.react-calendar__month-view__weekdays {
+    text-align: center;
+    text-transform: capitalize;
+    font-weight: 400;
 
-        abbr {
-            // т.к в _reboot.scss определены некрасивые штуки
-            cursor: auto;
-            text-decoration: none;
-        }
+    abbr {
+        // т.к в _reboot.scss определены некрасивые штуки
+        cursor: auto;
+        text-decoration: none;
     }
-    .react-calendar__month-view__weekdays__weekday {
-        padding: 0.3em;
-        background-color: $element-grey;
-        color: white;
-        padding-bottom: 0.25rem;
-        padding-top: 0.25rem;
-        border: 2px solid white;
-        @media (max-width: 576px) {
-            border-width: 1px;
-        }
-
-        border-top: 0;
-        white-space: pre !important;
-    }
-    .react-calendar__month-view__weekdays__weekday:first-of-type {
-        margin-left: 0;
-    }
-    .react-calendar__month-view__weekdays__weekday:last-of-type {
-        margin-right: 0;
+}
+.react-calendar__month-view__weekdays__weekday {
+    padding: 0.3em;
+    background-color: $element-grey;
+    color: white;
+    padding-bottom: 0.25rem;
+    padding-top: 0.25rem;
+    border: 2px solid white;
+    @media (max-width: 576px) {
+        border-width: 1px;
     }
 
-    .react-calendar__month-view__days__day--neighboringMonth {
-        color: #757575;
+    border-top: 0;
+    white-space: pre !important;
+}
+.react-calendar__month-view__weekdays__weekday:first-of-type {
+    margin-left: 0;
+}
+.react-calendar__month-view__weekdays__weekday:last-of-type {
+    margin-right: 0;
+}
+
+.react-calendar__month-view__days__day--neighboringMonth {
+    color: #757575;
+}
+
+.react-calendar__tile {
+    font-size: 1.25rem;
+    @media (max-width: 576px) {
+        font-size: 1rem;
     }
 
-    .react-calendar__tile {
-        font-size: 1.25rem;
-        @media (max-width: 576px) {
-            font-size: 1rem;
-        }
-
-        width: 100%;
-        border: 2px solid white;
-        @media (max-width: 576px) {
-            border-width: 1px;
-        }
-
-        text-align: center;
-        height: 64px;
-        white-space: nowrap;
-        padding: 0.4em 0.25em;
-        color: white;
-        background: $element-grey;
-        transition: background-color 0.3s;
-        font-weight: 500;
-    }
-    .react-calendar__tile:disabled {
-        background-color: #8a8a8a !important;
-    }
-    .react-calendar__tile:enabled:hover,
-    .react-calendar__tile:enabled:focus {
-        background-color: #e6e6e6 !important;
-    }
-    .react-calendar__tile--active {
-        background-color: #2a87bd !important;
-    }
-    .react-calendar__tile--active:enabled {
-        background: $primary !important;
-    }
-    .react-calendar__tile--active:enabled:hover,
-    .react-calendar__tile--active:enabled:focus {
-        background: $primary !important;
-    }
-    .react-calendar--selectRange .react-calendar__tile--hover {
-        background-color: #e6e6e6 !important;
+    width: 100%;
+    border: 2px solid white;
+    @media (max-width: 576px) {
+        border-width: 1px;
     }
 
-    .react-calendar__month-view__days__day--weekend {
-        background-color: #bcbcbc;
-    }
+    text-align: center;
+    height: 64px;
+    white-space: nowrap;
+    padding: 0.4em 0.25em;
+    color: white;
+    background: $element-grey;
+    transition: background-color 0.3s;
+    font-weight: 500;
+}
+.react-calendar__tile:disabled {
+    background-color: #8a8a8a !important;
+}
+.react-calendar__tile:enabled:hover,
+.react-calendar__tile:enabled:focus {
+    background-color: #e6e6e6 !important;
+}
+.react-calendar__tile--active {
+    background-color: #2a87bd !important;
+}
+.react-calendar__tile--active:enabled {
+    background: $primary !important;
+}
+.react-calendar__tile--active:enabled:hover,
+.react-calendar__tile--active:enabled:focus {
+    background: $primary !important;
+}
+.react-calendar--selectRange .react-calendar__tile--hover {
+    background-color: #e6e6e6 !important;
+}
+
+.react-calendar__month-view__days__day--weekend {
+    background-color: #bcbcbc;
 }
