+++ src/common/components/ContentBox/ContentBox.module.scss
@@ -1,4 +1,4 @@
-@import '@/styles/variables';
+@import '@/common/styles/variables';
 
 .wrapper {
     :global {
