+++ src/common/routes/ghostRoutes.jsx
