+++ src/common/styles/index.scss
@@ -1,5 +1,6 @@
 @import './variables';
 @import '~bootstrap/scss/bootstrap';
+@import '~@brainhubeu/react-carousel/lib/style.css';
 
 /* Webkit должен рендерить кнопки так, как у нас */
 .btn {
