+++ src/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import { Provider } from 'react-redux'
-import store from '@/store/store'
+import store from '@/common/store/store'
 import App from './App'
 
 ReactDOM.render(
