+++ src/modules/auth/login/Login.module.scss
@@ -1,4 +1,4 @@
-@import '@/styles/variables';
+@import '@/common/styles/variables';
 
 .loginPage {
     height: 80vh;
