+++ src/modules/auth/signup/Signup.module.scss
@@ -1,4 +1,4 @@
-@import '@/styles/variables';
+@import '@/common/styles/variables';
 
 .background {
     min-height: 100vh;
