+++ src/modules/landing/Landing.module.scss
@@ -1,4 +1,4 @@
-@import '@/styles/variables';
+@import '@/common/styles/variables';
 
 .landingNumbers {
     font-family: Quantico, sans-serif;
