+++ src/modules/queue/components/SlotCalendar/SlotCalendar.module.scss
@@ -1,4 +1,4 @@
-@import '@/styles/variables';
+@import '@/common/styles/variables';
 
 .wrapper {
     :global {
