+++ src/pages/_app.jsx
@@ -5,7 +5,7 @@ import { Provider } from 'react-redux'
 import Header from '@/common/components/Header/Header'
 import Footer from '@/common/components/Footer/Footer'
 
-import store from '@/store/store'
+import store from '@/common/store/store'
 
 import '@fontsource/roboto'
 import '@fontsource/roboto/300.css'
@@ -14,7 +14,7 @@ import '@fontsource/roboto/900.css'
 import '@fontsource/quantico'
 import '@fontsource/open-sans'
 import '@fontsource/open-sans/300.css'
-import '@/styles/index.scss'
+import '@/common/styles/index.scss'
 
 const MyApp = ({ Component, pageProps }) => (
     <>
